fix compiler warning

+++ optimize.c
@@ -2200,9 +2200,9 @@ opt_init(opt_state_t *opt_state, struct icode *ic)
 	/*
 	 * Overflow check.
 	 */
-	block_memsize = 2 * opt_state->n_blocks * opt_state->nodewords * sizeof(*opt_state->space);
-	edge_memsize = opt_state->n_edges * opt_state->edgewords * sizeof(*opt_state->space);
-	if (opt_state->n_edges && edge_memsize / sizeof(*opt_state->space) / opt_state->n_edges != opt_state->edgewords) {
+	block_memsize = (size_t)2 * opt_state->n_blocks * opt_state->nodewords * sizeof(*opt_state->space);
+	edge_memsize = (size_t)opt_state->n_edges * opt_state->edgewords * sizeof(*opt_state->space);
+	if (opt_state->n_edges && edge_memsize / sizeof(*opt_state->space) / opt_state->n_edges != (size_t)opt_state->edgewords) {
 		opt_error(opt_state, "integer multiply overflow");
 	}
 	if (block_memsize > SIZE_MAX - edge_memsize) {
